tb_detection_app/requirements.txt
flask>=3.0
flask-cors>=4.0
tensorflow>=2.13
numpy>=1.24
pillow>=10.0
opencv-python-headless>=4.8
